Prevent hang during daemon stop operation (when daemon thread are in blocking state)
Previously when importing xmlrpc module, we add a call to socket.setdefaulttimeout(5)
So every created socket without specific call to settimeout() or setblocking() inherits a timeout value of 5.

Because of lib remapping, and imports cleanup this setting is not done on daemon startup.
This may lead to blocking socket operations that were previously non blocking.

Possible side effect is that deamon threads stop doesn't terminate.
=> systemd may kill opensvc.daemon when stop time is exceeded

Daemon stop waits for all its threads, but some doesn't stop.

For example, if 'om mon -w' process is running, there is an associated client handler thread that is running
listener:ClientHandler.handle_h2_client()
    ...
    while True:
        if self.stopped():
            break
        try:
            data = self.tls_conn.recv(65535)  <- was now blocking until client connexion ends