Allow the keywords to define a custom DEFAULT fallback keyword
For example sync#*.schedule falls back to DEFAULT.sync_schedule, but without
the required infrastructure in the Extconfig class, this custom fallback was
done by the caller of oget().

This patch also fixes the "leaf > head" oget() codepath for unscoped keywords.

And adds a DEFAULT.run_schedule default_keyword for task#*.schedule, with a
None default value.