Don't validate global resource keyword value from encap
The scoping might not allow the encap to get a value.
For example:

[disk#0]
type = zpool
name@sol114-1 = data-c22svc31a
name@sol114-2 = data-c22svc31b

where sol114-* are global nodes.

This configuration produces a validation error in the encapnodes because
the "name" keyword is mandatory but the encapnodes don't have a scoped value.

This patch also make the global nodes to impersonate as an encapnode to
validate the encap resources keyword values.