Fix undue tasks "task is a already running" errors
The "get_running"-intent lock can prevent the tasks to acquire the "run"-intent
lock (timeout=0), which causes an undue "task is a already running" error and
the task schedule to fail a run that should have run.

Implement a lockless get_running(), making use of the new garanty that the
lockfile is truncated on release, introduced by the previous patch.

This implementation is also more efficient.