Don't ask the daemon to resolve the boot action object selector
The daemon might not be ready to resolve yet, and the objects selected are
always local.

This fixes a "no match" case on initial boot action, causing the pre-boot
status.json to be loaded instead of refreshed ones (in turn causes the
restart=<n> resource to be unduly restarted on service instances that were
up before the boot).