Add the "om node drain" command and orchestration
"om node drain" returns immediately, while the orchestration is
in progress.

"om node drain --wait" returns when the node is drained. This
command can be run if a drain is already in progress.

The drain orchestration is similar to the shutdown orchestration,
but the node is frozen to prevent instances restart.

To return to normal service, just unfreeze the node.