Don't purge the status.json on action
We don't use it anyway, so do not provoke an inline "status -r" in the
monitor thread when it needs reading the status.json but cannot find it.

Also remove the lock acquire from the status.json load codepath. Now that
this file is not longer overwritten, this lock is useless.