Set use-vc in resolv.conf options for containers using backend ipaddr
When the first namespace is not available the glibc udp dns request will
have to wait until timeout*attemps before trying the next nameserver, which
is 10s by default and 1s minimum (timeout:1 attempts:1).

Setting the use-vc option forces a tcp request instead of a udp request.
A tcp request will fail immediately if the nameserver is down.

Note: alpine's musl libc is not affected by this issue as it requests all
defined nameservers and returns the first response. But setting use-vc there
too is ok.