Fix two potential dns db desync situation
1/
The db cache was indexed by sorted dataset generation numbers (one per node).
So if n1 is at gen 0 and n2 at gen 1, then at next request n2 is at gen 0 and
n1 at gen 1, the same cache key (0, 1) is computed and the dns entries not
recomputed. (this one is far fetched, but doesn't hurt to close the window).

This patch orders the generation numbers by nodename.

2/
The db cache key was computed on set (after entries computation), so if the
dataset changed between the start of computation and the cache set, these
changes would be reflected only after another dataset change, which might be
10 minutes later at most.

This patch compute the cache key before computing the dns db entries, and pass
that key to the cache_set() method. cache_set() no longer computes the key
itself.