Fix a resource info corruption on peer node labels change
A consequence of this corruption for ip resource, which resource info contains
the ipaddr used in the dns db, the service dns records are dropped until the
next status refresh (may be up to 10m).

On peer node label changes the daemon triggers a refresh=False rewrite of the
services status.json, so all keys depending on labels are updated, but not
reevaluating the resources status. The resource dataset are fetched from the
resource status.last cache, which did not include the "info" key, causing a
loss of information.

This patch add the "info" key to status.last, and restore this cached
information on print_status_data_eval(refresh=False,...).