Support registry creds config as secrets
A new container.docker and container.podman keyword "registry_creds" can be
set in the resource or/and in DEFAULT to the name of a secret in the same
namespace.

If this secret contains a config.json key, which value is the content of
the config.json generated by a "docker login <registry dnsname>".

For example, the creds secret can be created by:

om test/sec/creds-registry-opensvc-com create
om test/sec/creds-registry-opensvc-com add --from ~/.docker/config.json

Upon container start and pull, the appropriate command option is added to
point an up-to-date projection of the config.json.

With podman --authfile, with docker --config.