Accept a timeout in Node::_daemon_unlock() and use a 120s timeout in disk.drbd
A socket.timeout exception was raised after a default 5s if the handler did
not return yet. The handler would try to return in 5s too, but can return a
few millisecs too late, and thus trigger the socket.timeout.

This patch allows the caller to specify a custom timeout in the request options,
and the request socket timeout is set 5s longer than the requested timeout.

The "om daemon lock release" now accepts a --timeout <duration> option.