Fix tcp ~15s hangs in containers tcp connections
When the device with the lowest mac is removed from the bridge or when
a new device with the lowest mac is added to the bridge, all containers
can experience tcp hangs while the arp table resynchronizes.

Setting a mac address to the bridge explicitely avoids these mac address
changes.

Forge the mac address using a 0a:58 prefix followed by the bridge ipv4
address converted to hexa (same algorithm used in k8s).

Plug the mac addr setting in the "om net setup" codepath. Incrementally
configure it, so it will be set upon upgrade for bridge created before
this patch.

Example:

root@demo2:~# om net set
@ n:demo2
  ip link set dev obr_net1 address 0a:58:0a:21:04:01
  ip route replace 10.33.0.0/22 via 5.196.34.153 table main
  ip route replace 10.33.4.0/22 dev obr_net1 table main