Move pool volume_env parsing from volume resource to the base pool class
This pool internals had no place to be in the resource driver.

Split _configure_volume() out of configure_volume() in the volume resource
driver, so _configure_volume() can be used to factor a volatile volume
object.

Finally, make the volsvc lazy return a no-cf volatile object when the
persistent object is not created yet.

This change makes it possible to resolve the volume#<n>.mnt reference even
if the volume object does not exist yet.