Add alternate 'reboot' split action when split brain cluster node has no quorum
When power-on feature is not easy to use, it may be simpler to reboot node instead of crash node.

New kw added: cluster.split_action
It can be defined in node.conf or cluster.conf, and can have 'crash' or 'reboot' value.
When undefined, 'crash' value it used.

When defined to 'reboot', a reboot is called instead of default 'crash' when split cluster doesn't get enough
votes during split brain situation.