Fix a resource info clear issue at the begining of tasks run action
This is an issue because the ip info contains the ipaddr exposed via the
cluster dns. So the ip resource could lose the resolution on run actions
if a resolution query is submitted during the run action.

This patch moves the Resource::status_info() call out of Svc::status_data_eval()
to Resource::status(), so everytime status() refreshes the main status data it
also updates the status_info() and updates the cache. Which garanty the
last_status_data cache file contains a valid "info" section.