Remove useless DEFAULT_STATUS_GROUP
Every object has a kwstore attribute now, which is an iterable returning
known drvgroup names.

No need to duplicate that iterable.

Move it to the only valid user, which is the instance renderer, which still
needs to be told how to order the drvgroup names, as it has no access to the
kwstore.