Add second-granularity scheduling
The timerange can now be expressed as HH:MM:SS.
The HH:MM format is interpreted as HH:MM:00.

Add a Schedule class implementing an iterator on next valid
datetimes, given a last run time (default now) and a schedule
expression.

Use this iterator in the daemon scheduler thread to improve
cpu usage, waking up frequently only to janitor spawned processes,
and on demand (less frequently) refill the future run slots for
terminated processes.

Now that determining the next valid date of a scheduled task is
fast, remove the --verbose option of "print schedule" and always
display this information.