Make the scheduler fork the object action
Instead of the Mgr() object, which shortcuts a lot of uneeded
optparsing and object selection code.

Stop coalescing the same action on multiple objects, and fork
one process for each object action, using the same session_id
for command cache sharing.

Purge the session cache when the last forked process using it
terminates.