Add a {uri_ip:...} reference modifier
Embedding an ipv6 in a listener like with {dns[0]}:5300 produces an
ambiguous definition (example ::1:5300).

Most product (including wget and pdns) expect the ipv6 inside brackets
(example [::1]:5300).

Add the "uri_ip" modifier so {uri_ip:dns[0]}:5300 evaluates as

* 127.0.0.1:5300 if dns = ["127.0.0.1"]
* [::1]:5300 if dns = ["::1"]

This will be used in the system/svc/dns template to abstract the
definition on ipv4 or ipv6 setups.