Fix a "dict changed size during iteration" in JournaledData::_set_lk(key, val)
The current value at <key> is compared to the new value passed as argument
via a recursive iterator.

The new value was deepcopied before store, but after comparison, and thus could
change during iteration.

This patch moves the deepcopy from _set() to _set_lk(), ie before comparison.

The patch fragment merge also called _set(), and now the patch value is stored
without deepcopy. This is a valid co-lateral optimization, as nothing changes
the patches on the applier endpoint.