Allow shared volumes from non shared-capable pool on mono-node clusters
When testing a service template with shared=true volumes on a mono-node
clusters, chances are opensvc won't find a shared-capable pool.

Implement a fallback on non shared-capable pool if a shared-capable pool
could not be found in this case.

Log a warning that this volume will not be able to failover if nodes get
added later on.