Add a DEFAULT.priority keyword to the Svc objects configuration
The priority is an integer defaulting to 100.
The lower the number the higher the priority.

The monitor thread now runs orchestration on Svc objects ordered by ascending
priority, so higher priority Svc are granted a slot in the max_parallel
constrained subprocess list.

A new "prioritizer" RBAC role is added to allow a non-root user to
set the priority of a Svc.

The priority shows in the print status dataset and its human readable format
(when not aligned with the default value).