Multi-claims volume fixes
Remove the patchs trying to workaround vol provision issues when access is not
set to r.x by a flex consumer: Now the access is forced to ..x.

Also separate the rollback boolean in 2 booleans :
* one for the vol start rollbacking
* one for the volume resource state flag

So even if we didn't start the vol instance we can rollback the state flag.