Ignore some rmtree error on lxc unprovision's cleanup of the rootfs
The rootfs might contain busy mount points, handled by a fs resource in the
svc or in a vol.

In this case, the rmtree(<rootfs>) will raise an exception on the first error,
exiting the cleanup too early.

This patch adds a error handler discarding EBUSY and ENOTEMPTY, so the cleanup
actually clean all cleanable files.