Add a exposed_devs() method to disk.zpool
As fs.zfs reports its "<pool name>/<dataset name>" as the exposed devs,
the disk.zpool should report is "<pool name>".

This will allow a volume with a disk.pool resource to have a valid "device"
property, that we can use to identify the volume providing the pool
hosting a fs.zfs resource in a svc object.