Destroy a fs.zfs dataset on unprovision even if the zpool is exported
Import the dataset temporarily to destroy the dataset if its not already
imported.

If the zpool is provided by a volume resource, also temporarily start it.

The "unprovision" orchestration stops the service, so this situation is
quite common. And a svc with a container.lxc with os datasets (/var, /home,
...) needs its rootfs really emptied to be able to reprovision.