Don't clear the "start failed" state while an orchestration is in progress
We checked only that the local instance's global_expect value was None to
allow clearing a "start failed" state, which caused a ping pong on 'toc'
action for services configured with monitor_action=switch.

This orchestration relies on the "start failed" to remain until the "placed"
global expect is cleared, because if cleared the optimal instance can become
the cleared one, causing a failback.