Use the start_timeout keyword in the app resource drivers
Use it as a timeout for a new wait loop in the start codepath.

When the app resource start script has run successfully, the check
script might still evaluate a down state (ex mysqld socket not ready yet,
and a checker using a mysql \q).

In this situation the instance startup reports a successful start but a not-up
end state.

This patch adds a loop waiting for the app state to become up after a successful
starter execution.