Remove an undue status log warning in the disk.md driver
on the failover node of a shared disk.md resource, the driver would warn
about a "unable to find a devpath" and report a "down" status.

This is not the expected behaviour. Not being able to find a devpath is
normal and does not mandates a warn-level status log.

this warn was unduly introduced by on april 16th 2021.