Fix a 3PARdata scsi pr oddity
Some arrays, like 3PARdata and Clariion, overwrite the registration of a path
with the registration of its failover path.

On a disk accessible via 4 paths (from 2 hba to 2 controlers), after registering
all 4 paths only 2 registrations are reported by "sg_persist -k" instead of 4.

It is not fair to ask the admins to set a keyword to describe the number of
paths per registration to use.

This patch implements:

* on resource start (if not already started), the paths per registration of
each device is computed and stored in a persistant cache file.
  - if unchanged, the cache file is not updated
  - the value is <pr gen delta>/<visible registration>
  - the cache is a map indexed by disk_id (not devpath, which is not stable)

* on resource check, load the cache and use the disk's pp_reg factor to
determine what to expect.

* if the cache file are purged, an info status_log is displayed, explaining
why this particular check is disabled

(cherry picked from commit 23006d7544ab6c9ed3fd35638963b16a68986ce7)