Fix wrong resource status after start for resource in a // subset
For // subsets, resource actions are run in separate process.

If an action unset lazy props, this unset is not effective in the parent
process, so at the end of the action, when the parent process computes the
post-action instance status, the lingering lazy props can produce invalid
status.

This issue happens reliably with docker containers in a // subset.

This patch detects the presence of a normalized cache cleanup method in the
resource, and calls that after the // process joins, .. for state changing
action only.