Fix "print config --eval" error on unknown DEFAULT.start_timeout keyword
A lot of keywords are not declared explicitely in the DEFAULT section,
because they can be defined here just as default value for a
in-resource keyword.

The "eval --kw" is able to return the value of those without error, but
"print config --eval" was not.

This patch also adds a deprecation warning for keywords, and considers
deprecated keywords are still usable.