Fix node.oci evaluating as "None" instead of either docker or podman
If node.dbopensvc=None in the node.conf, chances are that this kw is
evaluated before node.oci during object actions.

In this case, the "None" value is stored in the dereference cache with
key="...None". Then when node.oci is evaluated the same key is formatted
and "None" is returned from the cache instead of the expected None
(NoneType).

This patch implements a shortcut for dereferencing None. In this case
alway return None, without looking at the cache.