Allow using the disk.raw driver when the kmod is not installed
Add a capability drivers.resource.disk.raw.cdev if the raw kmod and
the raw command are installed.

Condition using the raw command to this cap.

With this patch, the disk.raw driver can be used only for inventoring
bdevs with "pushdisks", even when the system is not raw-capable. The
create_char_devices=false must be set in this case, to avoid stop and
start errors.