Fix scheduler job failures when many rids are batched together
The scheduler asks the CRM to create a file so it can verify the CRM has
really started working.

The filename is forged by the scheduler and passed to the CRM via the
process environment.

The filename format the scheduler uses produces a variable length
string, due to the list of rids being embedded. Many filesystems have a
255 bytes limit for filename, so the filename could cause
core/scheduler.py.timestamp() to fail when creating the file.

Use a simple "launched-<uuid>-<sid>" format instead.