Fix undue daemon start abort on pid conflict
The first fork() of the double-fork daemonization of the agent daemon returned instead of exiting immediately.
This caused the parent process to linger in defunct state. This defunct process confuses the daemon abort test
into thinking the pid read from the pidfile is still running if by chance the child pid is equals to the
parent (defunct) pid.

Add delay from the second fork reduce the risk to find running main, and fork1 process.

## Reproducer:

Force pid `<i>` in pidfile.
Start the daemon

    PID  DESC
    i    main => fork() <defunct>
    i+1    fork1 => fork() <defunct>
    i+2      fork2 => read <i> from pidfile, check if <i> is alive => abort

## With `os._exit()` in the forked parent instead of `return` and with delay in fork2:

    PID  DESC
    i    main => fork() <reaped>
    i+1    fork1 => fork() <reaped>
    i+2      fork2 => read <i> from pidfile, check if <i> is alive => pass